backend/pyproject.toml
[project]
name = "etl-backend"
version = "0.1.0"
description = "FastAPI backend for ETL Studio"
requires-python = ">=3.12"
dependencies = [
    "app-shared",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "boto3>=1.35.0",
    "celery[redis]>=5.4.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.49b0",
]

[project.optional-dependencies]
dev = ["ruff>=0.8.0", "mypy>=1.11.0", "httpx>=0.27.0"]

[tool.uv.sources]
app-shared = { workspace = true }

[tool.uv.scripts]
dev = "uvicorn app.main:app --reload --port 8000"
migrate = "alembic upgrade head"
revision = "alembic revision --autogenerate -m"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
